feat: Move where clause meethods to `org.spin.db.WhereClauseUtils`.
